Use functions instead of classes
