+++ packages/framework/tree-agent/src/methodBinding.ts
@@ -7,8 +7,7 @@
 import type { TreeNodeSchema, TreeNodeSchemaClass } from "@fluidframework/tree";
 import { NodeKind } from "@fluidframework/tree";
 import type { z } from "zod";
-
-import { instanceOf } from "./utils.js";
+import { instanceOf } from "./renderZodTypeScript.js";
 
 /**
  * A utility type that extracts the method keys from a given type.
