+++ packages/framework/tree-agent/src/utils.ts
@@ -3,13 +3,10 @@
  * Licensed under the MIT License.
  */
 
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
-
 import { assert } from "@fluidframework/core-utils/internal";
 import { isFluidHandle } from "@fluidframework/runtime-utils";
 import { UsageError } from "@fluidframework/telemetry-utils/internal";
-import type { ImplicitFieldSchema, TreeNodeSchemaClass } from "@fluidframework/tree";
+import type { ImplicitFieldSchema } from "@fluidframework/tree";
 import type {
 	InsertableContent,
 	TreeNode,
@@ -19,12 +16,10 @@ import type {
 import {
 	ArrayNodeSchema,
 	MapNodeSchema,
-	ObjectNodeSchema,
 	RecordNodeSchema,
 	TreeAlpha,
 } from "@fluidframework/tree/alpha";
 import { NodeKind, normalizeFieldSchema } from "@fluidframework/tree/internal";
-import { z } from "zod";
 
 /**
  * Subset of Map interface.
@@ -193,318 +188,6 @@ export function unqualifySchema(schemaIdentifier: string): string {
 	}
 	return matches[0];
 }
-
-/**
- * Returns the TypeScript source code corresponding to a Zod schema. The schema is supplied as an object where each
- * property provides a name for an associated Zod type. The return value is a string containing the TypeScript source
- * code corresponding to the schema. Each property of the schema object is emitted as a named `interface` or `type`
- * declaration for the associated type and is referenced by that name in the emitted type declarations. Other types
- * referenced in the schema are emitted in their structural form.
- * @param schema - A schema object where each property provides a name for an associated Zod type.
- * @param details - Optional details about the schema. The fields will be set according to the details in the given schema.
- * @returns The TypeScript source code corresponding to the schema.
- */
-export function formatZodType(type: z.ZodTypeAny): string {
-	const writer = new ZodTypeWriter();
-	writer.appendType(type);
-	return writer.toString();
-}
-
-class ZodTypeWriter {
-	private result = "";
-	private startOfLine = true;
-	private indent = 0;
-
-	public appendType(type: z.ZodTypeAny, minPrecedence = TypePrecedence.Object): void {
-		const shouldParenthesize = getTypePrecendece(type) < minPrecedence;
-		if (shouldParenthesize) {
-			this.append("(");
-		}
-		this.appendTypeDefinition(type);
-		if (shouldParenthesize) {
-			this.append(")");
-		}
-	}
-
-	public toString(): string {
-		return this.result;
-	}
-
-	private append(s: string): void {
-		if (this.startOfLine) {
-			this.result += "    ".repeat(this.indent);
-			this.startOfLine = false;
-		}
-		this.result += s;
-	}
-
-	private appendNewLine(): void {
-		this.append("\n");
-		this.startOfLine = true;
-	}
-
-	private appendTypeDefinition(type: z.ZodTypeAny): void {
-		switch (getTypeKind(type)) {
-			case z.ZodFirstPartyTypeKind.ZodString: {
-				this.append("string");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodNumber: {
-				this.append("number");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodBoolean: {
-				this.append("boolean");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodDate: {
-				this.append("Date");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodUndefined: {
-				this.append("undefined");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodNull: {
-				this.append("null");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodUnknown: {
-				this.append("unknown");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodArray: {
-				this.appendArrayType(type);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodObject: {
-				this.appendObjectType(type);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodUnion: {
-				this.appendUnionOrIntersectionTypes(
-					(type._def as z.ZodUnionDef).options,
-					TypePrecedence.Union,
-				);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodDiscriminatedUnion: {
-				this.appendUnionOrIntersectionTypes(
-					[...(type._def as z.ZodDiscriminatedUnionDef<string>).options.values()],
-					TypePrecedence.Union,
-				);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodIntersection: {
-				this.appendUnionOrIntersectionTypes(
-					[
-						(type._def as z.ZodIntersectionDef).left,
-						(type._def as z.ZodIntersectionDef).right,
-					],
-					TypePrecedence.Intersection,
-				);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodTuple: {
-				this.appendTupleType(type);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodRecord: {
-				this.appendRecordType(type);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodMap: {
-				this.appendMapType(type);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodLiteral: {
-				this.appendLiteral((type._def as z.ZodLiteralDef).value);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodEnum: {
-				this.append(
-					(type._def as z.ZodEnumDef).values.map((value) => JSON.stringify(value)).join(" | "),
-				);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodOptional: {
-				this.appendUnionOrIntersectionTypes(
-					[(type._def as z.ZodOptionalDef).innerType, z.undefined()],
-					TypePrecedence.Union,
-				);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodReadonly: {
-				this.appendReadonlyType(type);
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodEffects: {
-				const schema = instanceOfs.get(type);
-				if (schema === undefined) {
-					throw new UsageError(
-						`Unsupported zod effects type when formatting helper types: ${getTypeKind(type)}`,
-					);
-				}
-				this.append(getFriendlyName(schema));
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodVoid: {
-				this.append("void");
-				return;
-			}
-			case z.ZodFirstPartyTypeKind.ZodLazy: {
-				this.appendType((type._def as z.ZodLazyDef).getter());
-				return;
-			}
-			default: {
-				throw new UsageError(
-					`Unsupported type when formatting helper types: ${getTypeKind(type)}`,
-				);
-			}
-		}
-	}
-
-	private appendArrayType(arrayType: z.ZodTypeAny): void {
-		this.appendType((arrayType._def as z.ZodArrayDef).type, TypePrecedence.Object);
-		this.append("[]");
-	}
-
-	private appendObjectType(objectType: z.ZodTypeAny): void {
-		this.append("{");
-		this.appendNewLine();
-		this.indent++;
-		// eslint-disable-next-line prefer-const
-		for (let [name, propertyType] of Object.entries(
-			(objectType._def as z.ZodObjectDef).shape(),
-		)) {
-			this.append(name);
-			if (getTypeKind(propertyType) === z.ZodFirstPartyTypeKind.ZodOptional) {
-				this.append("?");
-				propertyType = (propertyType._def as z.ZodOptionalDef).innerType;
-			}
-			this.append(": ");
-			this.appendType(propertyType);
-			this.append(";");
-			this.appendNewLine();
-		}
-		this.indent--;
-		this.append("}");
-	}
-
-	private appendUnionOrIntersectionTypes(
-		types: readonly z.ZodTypeAny[],
-		minPrecedence: TypePrecedence,
-	): void {
-		let first = true;
-		for (const type of types) {
-			if (!first) {
-				this.append(minPrecedence === TypePrecedence.Intersection ? " & " : " | ");
-			}
-			this.appendType(type, minPrecedence);
-			first = false;
-		}
-	}
-
-	private appendTupleType(tupleType: z.ZodTypeAny): void {
-		this.append("[");
-		let first = true;
-		for (const type of (tupleType._def as z.ZodTupleDef<z.ZodTupleItems, z.ZodType>).items) {
-			if (!first) {
-				this.append(", ");
-			}
-			if (getTypeKind(type) === z.ZodFirstPartyTypeKind.ZodOptional) {
-				this.appendType((type._def as z.ZodOptionalDef).innerType, TypePrecedence.Object);
-				this.append("?");
-			} else {
-				this.appendType(type);
-			}
-			first = false;
-		}
-		const rest = (tupleType._def as z.ZodTupleDef<z.ZodTupleItems, z.ZodType | null>).rest;
-		if (rest !== null) {
-			if (!first) {
-				this.append(", ");
-			}
-			this.append("...");
-			this.appendType(rest, TypePrecedence.Object);
-			this.append("[]");
-		}
-		this.append("]");
-	}
-
-	private appendRecordType(recordType: z.ZodTypeAny): void {
-		this.append("Record<");
-		this.appendType((recordType._def as z.ZodRecordDef).keyType);
-		this.append(", ");
-		this.appendType((recordType._def as z.ZodRecordDef).valueType);
-		this.append(">");
-	}
-
-	private appendMapType(mapType: z.ZodTypeAny): void {
-		this.append("Map<");
-		this.appendType((mapType._def as z.ZodMapDef).keyType);
-		this.append(", ");
-		this.appendType((mapType._def as z.ZodMapDef).valueType);
-		this.append(">");
-	}
-
-	private appendLiteral(value: unknown): void {
-		this.append(
-			typeof value === "string" || typeof value === "number" || typeof value === "boolean"
-				? JSON.stringify(value)
-				: "any",
-		);
-	}
-
-	private appendReadonlyType(readonlyType: z.ZodType): void {
-		this.append("Readonly<");
-		this.appendType((readonlyType._def as z.ZodReadonlyDef).innerType);
-		this.append(">");
-	}
-}
-
-function getTypeKind(type: z.ZodType): z.ZodFirstPartyTypeKind {
-	return (type._def as z.ZodTypeDef & { typeName: z.ZodFirstPartyTypeKind }).typeName;
-}
-
-const enum TypePrecedence {
-	Union = 0,
-	Intersection = 1,
-	Object = 2,
-}
-
-function getTypePrecendece(type: z.ZodType): TypePrecedence {
-	switch (getTypeKind(type)) {
-		case z.ZodFirstPartyTypeKind.ZodEnum:
-		case z.ZodFirstPartyTypeKind.ZodUnion:
-		case z.ZodFirstPartyTypeKind.ZodDiscriminatedUnion: {
-			return TypePrecedence.Union;
-		}
-		case z.ZodFirstPartyTypeKind.ZodIntersection: {
-			return TypePrecedence.Intersection;
-		}
-		default: {
-			return TypePrecedence.Object;
-		}
-	}
-}
-
-/**
- * Create a Zod schema for a SharedTree schema class.
- * @alpha
- */
-export function instanceOf<T extends TreeNodeSchemaClass>(
-	schema: T,
-): z.ZodType<InstanceType<T>, z.ZodTypeDef, InstanceType<T>> {
-	if (!(schema instanceof ObjectNodeSchema)) {
-		throw new UsageError(`${schema.identifier} must be an instance of ObjectNodeSchema.`);
-	}
-	const effect = z.instanceof(schema);
-	instanceOfs.set(effect, schema);
-	return effect;
-}
-
-const instanceOfs = new WeakMap<z.ZodTypeAny, ObjectNodeSchema>();
 
 /**
  * Adds all (optionally filtered) schemas reachable from the given schema to the given set.
